completed initial single precision timing test with ./main working

+++ src/MatrixClasses/Timer.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-float Timer::gtod(void) const
+double Timer::gtod(void) const
 {
     struct timeval tv;
     gettimeofday(&tv, (struct timezone*)nullptr);
@@ -11,9 +11,9 @@ float Timer::gtod(void) const
 
 void Timer::print(ostream& os) const
 {
-    float tmin, tmax, tavg;
+    double tmin, tmax, tavg;
 
-    float treal = total_real_;
+    double treal = total_real_;
 
     // get rank
     int mype;
@@ -23,10 +23,10 @@ void Timer::print(ostream& os) const
     MPI_Comm_size(comm_, &npes);
 
     // get min and max real time
-    MPI_Allreduce(&treal, &tmin, 1, MPI_FLOAT, MPI_MIN, comm_);
-    MPI_Allreduce(&treal, &tmax, 1, MPI_FLOAT, MPI_MAX, comm_);
-    MPI_Allreduce(&treal, &tavg, 1, MPI_FLOAT, MPI_SUM, comm_);
-    tavg /= (float)(npes);
+    MPI_Allreduce(&treal, &tmin, 1, MPI_DOUBLE, MPI_MIN, comm_);
+    MPI_Allreduce(&treal, &tmax, 1, MPI_DOUBLE, MPI_MAX, comm_);
+    MPI_Allreduce(&treal, &tavg, 1, MPI_DOUBLE, MPI_SUM, comm_);
+    tavg /= (double)(npes);
 
     // get min and max number of calls
     int numcalls = ncalls_;
@@ -34,7 +34,7 @@ void Timer::print(ostream& os) const
     MPI_Allreduce(&numcalls, &nmin, 1, MPI_INT, MPI_MIN, comm_);
     MPI_Allreduce(&numcalls, &nmax, 1, MPI_INT, MPI_MAX, comm_);
     MPI_Allreduce(&numcalls, &navg, 1, MPI_INT, MPI_SUM, comm_);
-    navg = (float)(navg) / (float)(npes);
+    navg = (double)(navg) / (double)(npes);
 
     if (nmax == 0) return;
